# docs/swagger.yaml
basePath: /
definitions:
...
  models.ReleaseEntry:
    properties:
      anime_slug:
        type: string
      cover_url:
        type: string
      genres:
        items:
          type: string
        type: array
      release_time:
        type: string
      score:
        type: string
      title:
        type: string
      type:
        type: string
      url:
        type: string
    type: object
  models.ReleaseScheduleResponse:
    properties:
      confidence_score:
        type: number
      data:
        additionalProperties:
          items:
            $ref: '#/definitions/models.ReleaseEntry'
          type: array
        type: object
      message:
        type: string
      source:
        type: string
    type: object
  models.ScheduleByDayResponse:
    properties:
      confidence_score:
        type: number
      data:
        items:
          $ref: '#/definitions/models.ScheduleEntry'
        type: array
      message:
        type: string
      source:
        type: string
    type: object
  models.ScheduleEntry:
    properties:
      anime_slug:
        type: string
      cover_url:
        type: string
      genres:
        items:
          type: string
        type: array
      release_time:
        type: string
      score:
        type: string
      title:
        type: string
      type:
        type: string
      url:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API untuk scraping data drama Korea dari dramaqu.ad
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: DramaQu API
  version: "1.0"
paths:
...
  /api/v1/jadwal-rilis:
    get:
      consumes:
      - application/json
      description: Mengambil jadwal rilis anime per hari
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ReleaseScheduleResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get jadwal rilis
      tags:
      - jadwal-rilis
  /api/v1/jadwal-rilis/{day}:
    get:
      consumes:
      - application/json
      description: Mengambil jadwal rilis anime untuk hari tertentu
      parameters:
      - description: Nama hari (monday, tuesday, wednesday, thursday, friday, saturday,
          sunday)
        in: path
        name: day
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ScheduleByDayResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get jadwal rilis by day
      tags:
      - jadwal-rilis
...
swagger: "2.0"